Broken implementation of method calls [ci skip]
It's broken probably we're happily propagating NoMatch error results
when we actually did consume some input.

I propose changing the convention: now instead of a special NoMatch
constructor, we'll indicate current position in the error, and try_ will
check that against current position. If it is the same, means we didn't
consume any input.